agentQ/requirements.txt
# agentQ/requirements.txt

# --- Core Dependencies ---
pulsar-client
fastavro
python-dotenv
pyyaml
pyignite
httpx
uvicorn
websockets
werkzeug
Jinja2

# --- Observability ---
structlog
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-requests

# --- Graph & Data ---
gremlinpython
duckdb
elasticsearch

# --- AI & Embeddings ---
sentence-transformers
torch
matplotlib

# --- Tooling & Linters ---
kubernetes
flake8
semgrep
watchfiles

# Quantum Computing Libraries
# qiskit>=0.45.0
# cirq>=1.3.0
# pennylane>=0.33.0
# qiskit-aer>=0.13.0
# qiskit-ibmq-provider>=0.20.0
# qiskit-optimization>=0.6.0
# qiskit-machine-learning>=0.7.0

# Neuromorphic Computing Libraries
# brian2>=2.5.0
# nengo>=3.2.0
# nest-simulator>=3.5.0
# snntorch>=0.7.0
# spikingjelly>=0.0.0.0.14

# Additional Scientific Computing
scipy>=1.11.0
scikit-learn>=1.3.0
networkx>=3.1.0
